made create components reusable
