+++ src/utils/global_consts.js
@@ -1,0 +1,15 @@
+export const CATEGORY_OPTIONS = [
+    {value:"food", label:"Food"},
+    {value:"entertainment", label:"Entertainment"},
+    {value:"music", label:"Music"},
+    {value:"games", label:"Games"},
+    {value:"tv", label:"TV"},
+    {value:"movies", label:"Movies"},
+    {value:"anime", label:"Anime"},
+    {value:"books", label:"Books"},
+    {value:"sports", label:"Sports"},
+    {value:"kpop", label:"Kpop"},
+    {value:"other", label:"Other"} 
+];
+
+export const FIRESTORE_COLLECTION_NAME = "all_games";
